# docs/swagger.yaml
basePath: /api
definitions:
...
  agent.ErrorResponse:
    properties:
      error:
        example: workload not found
        type: string
    type: object
...
  agent.RegistryAuth:
    properties:
      registry:
        description: Registry host, e.g. "ghcr.io"
        type: string
      token_ref:
        description: Env-store key holding the token/PAT
        type: string
      username:
        description: Registry username; defaults to "x-access-token" for PATs
        type: string
    type: object
...
  agent.TunnelRequest:
    properties:
      token:
        example: eyJhIjoiNjA2...
        type: string
    type: object
...
  agent.Workload:
    properties:
      allowed_nodes:
        description: 'Node whitelist: empty/["*"] = all, otherwise node names/IDs'
        items:
          type: string
        type: array
      autostart:
        description: Auto-start when Jetty starts up
        type: boolean
      compose:
        description: Default compose (used if no arch-specific)
        type: string
      compose_amd64:
        description: 'Optional: amd64-specific compose'
        type: string
      compose_arm64:
        description: 'Optional: arm64-specific compose'
        type: string
      ip:
        description: Service IP (routed via WARP)
        type: string
      name:
        description: DNS hostname
        type: string
      owner:
        description: Node HWID
        type: string
      registry_auth:
        allOf:
        - $ref: '#/definitions/agent.RegistryAuth'
        description: |-
          RegistryAuth optionally authenticates this workload's image pull
          against a private registry. The token itself is never stored here -
          TokenRef names a key in the encrypted env store whose value is the
          registry password/PAT. When nil, pulls run with no Jetty-managed
          credentials (public images, or whatever the host docker config
          already provides) - identical to pre-feature behavior.
      revive:
        description: Auto-failover to another node if owner dies
        type: boolean
      tags:
        description: |-
          Tags group workloads for filtering and bulk operations (start
          all tagged "media", export all tagged "prod", etc.). Free-form
          strings; validated with validTagPattern. Sorted+deduped on
          ingest so the wire form is canonical.
        items:
          type: string
        type: array
      version:
        description: Unix timestamp
        type: integer
    type: object
  agent.WorkloadRequest:
    properties:
      allowed_nodes:
        example:
        - node1
        - node2
        items:
          type: string
        type: array
      autostart:
        example: true
        type: boolean
      compose:
        example: |-
          services:
            web:
              image: nginx
        type: string
      compose_amd64:
        example: |-
          services:
            web:
              image: nginx:amd64
        type: string
      compose_arm64:
        example: |-
          services:
            web:
              image: nginx:arm64
        type: string
      ip:
        example: 10.100.0.50
        type: string
      name:
        example: nginx
        type: string
      registry_auth:
        $ref: '#/definitions/agent.RegistryAuth'
      revive:
        example: true
        type: boolean
    type: object
info:
  contact:
    name: Jetty Support
    url: https://github.com/ncwardell/jetty
  description: P2P Docker Compose orchestration with Cloudflare WARP mesh networking
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Jetty API
  version: 0.0.x
paths:
...
  /heartbeat:
    post:
      consumes:
      - application/json
      description: Node-to-node liveness used in tunnel mode, where peers cannot reach
        each other directly over WARP. Refreshes the sender's LastSeen. Peer API key
        required.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Peer liveness ping (internal)
      tags:
      - internal
...
  /peer-announce:
    post:
      consumes:
      - application/json
      description: Node-to-node. A peer reports its identity, WARP IP, version and
        architecture so the receiver can add or refresh it in the peer table. Peer
        API key required.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Announce a peer's current address and state (internal)
      tags:
      - internal
...
  /tunnel/sync:
    post:
      consumes:
      - application/json
      description: Node-to-node. Accepts the cluster CFToken from a peer and starts
        or stops cloudflared to match. A node that has opted out via DELETE /tunnel?scope=node
        accepts the token but stays stopped. Peer API key required.
      parameters:
      - description: Cluster tunnel token; empty string tears the tunnel down
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/agent.TunnelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Receive a tunnel token broadcast (internal)
      tags:
      - internal
...
  /workloads:
    get:
      description: Returns all workloads in the cluster, optionally filtered by node
      parameters:
      - description: Filter by node name or ID (use 'local' for this node only)
        in: query
        name: node
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/agent.Workload'
            type: array
      summary: List all workloads
      tags:
      - workloads
    post:
      consumes:
      - application/json
      description: Creates and deploys a new Docker Compose workload with a mesh IP
      parameters:
      - description: Workload configuration
        in: body
        name: workload
        required: true
        schema:
          $ref: '#/definitions/agent.WorkloadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/agent.Workload'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
        "409":
          description: Mesh IP already in use
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
        "500":
          description: Deployment failed
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Deploy a new workload
      tags:
      - workloads
...
schemes:
- https
- http
security:
- ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    description: API key for authentication (JETTY_SECRET)
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"